requirements.txt
pyrogram
TgCrypto
python-dotenv
hachoir
ffmpeg-python
aiohttp
motor
pytz
pyromod
psutil
Pillow
humanize